requirements.txt
# web framework & server
fastapi
uvicorn[standard]
python-multipart

# env and config
python-dotenv

# RAG API client
requests

# OpenAI SDK
openai>=1.0.0

# CLI helper for your scripts
fire

# PDF parsing for your indexer
pymupdf

# Llama‑Index for ingestion/indexing
llama-index[all]

# LangGraph core orchestration
langgraph

# (optional) ASCII graph layouts
grandalf
